first commit of refactor, not fully tested
