+++ src/flair/collapse_isoforms_precise.py
@@ -6,57 +6,53 @@ import argparse
 import math
 import os
 from multiprocessing import Pool
+from functools import partial
 
-parser = argparse.ArgumentParser(description='collapse parse options',
-			usage='python collapse_isoforms_precise.py -q <query.psl>/<query.bed> [options]')
-required = parser.add_argument_group('required named arguments')
-required.add_argument('-q', '--query', type=str, default='', required=True, action='store',
-	dest='q', help='BED12 or PSL file of aligned/corrected reads. PSL should end in .psl')
-parser.add_argument('-o', '--output', type=str, action='store',
-	dest='o', default='', help='specify output file, should agree with query file type')
-parser.add_argument('-w', '--window', default=100, type=int,
-	action='store', dest='w', help='window size for grouping TSS/TES (100)')
-parser.add_argument('-s', '--support', default=0.25, type=float, action='store',
-	dest='s', help='minimum proportion(s<1) or number of supporting reads(s>=1) for an isoform (0.1)')
-parser.add_argument('-f', '--gtf', default='', type=str,
-	action='store', dest='f', help='GTF annotation file for selecting annotated TSS/TES')
-parser.add_argument('-m', '--max_results', default=2, type=int, action='store', dest='m',
-	help='maximum number of novel TSS or TES picked per isoform unless --no_redundant is specified (2)')
-parser.add_argument('-t', '--threads', default=2, type=int,
-	action='store', dest='t', help='number of threads to use (2)')
-parser.add_argument('-n', '--no_redundant', default='none', action='store', dest='n',
-	help='For each unique splice junction chain, report options include: \
-	none: multiple best TSSs/TESs chosen for each unique set of splice junctions, see M; \
-	longest: TSS/TES chosen to maximize length; \
-	best_only: single best TSS/TES used in conjunction chosen; \
-	longest/best_only override max_results argument immediately before output \
-	resulting in one isoform per unique set of splice junctions (default: none)')
-parser.add_argument('-c', '--clean', default=False, action='store_true', dest='c',
-	help='Specify this to not append read support to the end of each entry (default: not specified)')
-parser.add_argument('-i', '--isoformtss', default=False, action='store_true', dest='i',
-	help='when specified, TSS/TES for each isoform will be determined from supporting reads \
-	for individual isoforms (default: not specified, determined at the gene level)')
-parser.add_argument('-nosplice', default='chrM', action='store', dest='nosplice',
-	help='Comma separated list of chromosomes that should not have spliced isoforms (default: chrM)')
-parser.add_argument('--quiet', default=False,
-	action='store_true', dest='quiet', help='suppress output to stderr')
-args = parser.parse_args()
+#TODO: test this code with tes and tss URGENT
+def main():
+	parser = argparse.ArgumentParser(description='collapse parse options',
+				usage='python collapse_isoforms_precise.py -q <query.psl>/<query.bed> [options]')
+	required = parser.add_argument_group('required named arguments')
+	required.add_argument('-q', '--query', type=str, required=True, 
+		help='BED12 or PSL file of aligned/corrected reads. PSL should end in .psl')
+	parser.add_argument('-o', '--output', type=str, 
+		help='specify output file, should agree with query file type')
+	parser.add_argument('-w', '--window', default=100, type=int,
+		 help='window size for grouping TSS/TES (100)')
+	parser.add_argument('-s', '--support', default=0.25, type=float, 
+		help='minimum proportion(s<1) or number of supporting reads(s>=1) for an isoform (0.1)')
+	parser.add_argument('-f', '--gtf', type=str,
+		 help='GTF annotation file for selecting annotated TSS/TES')
+	parser.add_argument('-m', '--max_results', default=2, type=int, 
+		help='maximum number of novel TSS or TES picked per isoform unless --no_redundant is specified (2)')
+	parser.add_argument('-t', '--threads', default=2, type=int,
+		 help='number of threads to use (2)')
+	parser.add_argument('-n', '--no_redundant', default='none',
+		help='For each unique splice junction chain, report options include: \
+		none: multiple best TSSs/TESs chosen for each unique set of splice junctions, see M; \
+		longest: TSS/TES chosen to maximize length; \
+		best_only: single best TSS/TES used in conjunction chosen; \
+		longest/best_only override max_results argument immediately before output \
+		resulting in one isoform per unique set of splice junctions (default: none)')
+	parser.add_argument('-c', '--clean', default=False, action='store_true',
+		help='Specify this to not append read support to the end of each entry (default: not specified)')
+	parser.add_argument('-i', '--isoformtss', default=False, action='store_true',
+		help='when specified, TSS/TES for each isoform will be determined from supporting reads \
+		for individual isoforms (default: not specified, determined at the gene level)')
+	parser.add_argument('--nosplice', default='chrM',
+		help='Comma separated list of chromosomes that should not have spliced isoforms (default: chrM)')
+	parser.add_argument('--quiet', default=False, action='store_true',
+		help='suppress output to stderr')
+	args = parser.parse_args()
 
-try:
-	max_results, window, minsupport, psl = args.m, args.w, args.s, open(args.q)
-	if args.f:
-		gtf = open(args.f)
-except:
-	sys.stderr.write('Make sure all files (query, GTF) have valid paths and can be opened\n')
-	sys.exit(1)
-
-bed = args.q[-3:].lower() != 'psl'
-if args.o:
-	if args.o[-3:].lower() != args.q[-3:].lower():
-		sys.stderr.write('Make sure input and output file extensions agree\n')
-		sys.exit(1)
-else:  # default output name
-	args.o = args.q[:-3]+'collapsed.bed' if bed else args.q[:-3]+'collapsed.psl'
+	isbed = True
+	if args.query[-3:].lower() == 'psl':
+		isbed = False
+	
+	# This should really be an object that gets passed (that is not named args)
+	collapse_isoforms_precise(queryfile=args.query, max_results=args.max_results, threads=args.threads, 
+		window=args.window, isoformtss=args.isoformtss, minsupport=args.support, gtfname=args.gtf, 
+		clean=args.clean, nosplice=args.nosplice, no_redundant=args.no_redundant, outputfname=args.output, isbed=True, quiet=args.quiet)
 
 
 def get_junctions(line):
@@ -87,8 +83,7 @@ def overlap(coord0, coord1, tol=0):
 		coord0, coord1 = coord1, coord0
 	isoverlap = coord0[0] <= coord1[0] and coord1[0] <= coord0[1] - tol
 	if isoverlap:
-		return isoverlap, (min(coord0[1], coord1[1]) - coord1[0])/(coord0[1] - coord0[0])# + \
-		#(coord0[1] - coord1[0])/(coord1[1] - coord1[0]) / 2
+		return isoverlap, (min(coord0[1], coord1[1]) - coord1[0])/(coord0[1] - coord0[0])
 	return isoverlap, 0
 
 
@@ -124,77 +119,7 @@ def bin_search(query, data):
 	return i
 
 
-# unused
-#def interval_insert(query, data):
-#	i = int(round(len(data)/2))
-#	lower, upper = 0, len(data)
-#	if not upper:
-#		return [query]
-#	while True:
-#		if data[i][0] <= query[0]:  # i is too low of an index
-#			lower = i
-#			i = int(math.floor((lower + upper)/2.))
-#			if i == lower:
-#				break
-#		elif data[i][0] >= query[0]:  # i is too high of an index
-#			upper = i
-#			i = int(math.floor((lower + upper)/2.))
-#			if i == upper:
-#				i = -1
-#				break
-#	data = data[:i+1] + [query] + data[i+1:]
-#	return data
-#
-#def add_se(sedict, tss, tes, line, sedictkeys, support=1):
-#	loci = {}  # altered loci
-#	c = bin_search((tss, tes), sedictkeys)
-#	if c != -1:  # found
-#		for coord in sedictkeys[c:]:  # coordinates of a locus of single-exon reads
-#			if overlap((tss, tes), coord):
-#				if tss not in sedict[coord]['tss']:
-#					sedict[coord]['tss'][tss] = 0
-#				if tss not in sedict[coord]['tss_tes']:
-#					sedict[coord]['tss_tes'][tss] = {}
-#				if tes not in sedict[coord]['tss_tes'][tss]:
-#					sedict[coord]['tss_tes'][tss][tes] = 0
-#				if not loci:
-#					sedict[coord]['tss'][tss] += support  # add tss/tes data to existing locus
-#					sedict[coord]['tss_tes'][tss][tes] += support
-#					newcoord = tss if tss < coord[0] else coord[0], tes if tes > coord[1] else coord[1]
-#					loci[newcoord] = sedict.pop(coord)
-#				else:
-#					newlocus = sedict.pop(coord)
-#					oldlocus = list(loci.keys())[0]  # previous locus the tss/tes overlapped with
-#					if tss in newlocus['tss']:
-#						newlocus['tss'][tss] += support
-#						if tes in newlocus['tss_tes'][tss]:
-#							newlocus['tss_tes'][tss][tes] += support
-#						else:
-#							newlocus['tss_tes'][tss][tes] = loci[oldlocus]['tss_tes'][tss][tes] + support
-#					loci[oldlocus]['tss_tes'].update(newlocus['tss_tes'])  # combine previous and current loci
-#					loci[oldlocus]['tss'].update(newlocus['tss'])
-#					newcoord = oldlocus[0] if oldlocus[0] < coord[0] else coord[0],\
-#						oldlocus[1] if oldlocus[1] > coord[1] else coord[1]  # combined locus name
-#					loci[newcoord] = loci.pop(oldlocus)  # add all info under combined locus name
-#			else:
-#				break
-#		sedict.update(loci)
-#		sedictkeys = sorted(sedict.keys(), key=lambda x: x[0])
-#	else:  # define new locus
-#		locus = (tss,tes)
-#		sedict[locus] = {}
-#		sedict[locus]['tss'] = {}
-#		sedict[locus]['tss'][tss] = support
-#		sedict[locus]['tss_tes'] = {}
-#		sedict[locus]['tss_tes'][tss] = {}
-#		sedict[locus]['tss_tes'][tss][tes] = support
-#		sedict[locus]['line'] = line
-#		sedict[locus]['bounds'] = (tss, tes)
-#		sedictkeys = interval_insert((tss,tes),sedictkeys)
-#	return sedict, sedictkeys
-#
-
-def iterative_add_se(sedict, chrom, group, se):
+def iterative_add_se(sedict, chrom, group, se, all_se_by_chrom, window, isbed=True):
 	tss, tes = se[0], se[1]
 	support = all_se_by_chrom[chrom][se]['count']
 	if group not in sedict[chrom]:
@@ -213,7 +138,7 @@ def iterative_add_se(sedict, chrom, group, se):
 		sedict[chrom][group]['bounds'][1] = sedict[chrom][group]['bounds'][1] + \
 			support*(tes - sedict[chrom][group]['bounds'][1])/sedict[chrom][group]['bounds'][2]
 
-	strand = all_se_by_chrom[chrom][se]['line'][5] if bed else all_se_by_chrom[chrom][se]['line'][8]
+	strand = all_se_by_chrom[chrom][se]['line'][5] if isbed else all_se_by_chrom[chrom][se]['line'][8]
 	if strand not in sedict[chrom][group]['strand']:
 		sedict[chrom][group]['strand'][strand] = 0
 	sedict[chrom][group]['strand'][strand] += 1
@@ -229,14 +154,15 @@ def iterative_add_se(sedict, chrom, group, se):
 	return sedict, True
 
 
-def run_iterative_add_se(chrom):  # add single exon genes iteratively, assumes that reads are sorted
+def run_iterative_add_se(chrom, all_se_by_chrom, window, isbed):  # add single exon genes iteratively, assumes that reads are sorted
 	group = 0  # group number
 	sedict = {}
 	sedict[chrom] = {}
 	se_ordered = all_se_by_chrom[chrom].keys()
 	se_ordered = sorted(se_ordered, key=lambda x: x[1])
 	se_ordered = sorted(se_ordered, key=lambda x: x[0])
-	sedict, added = iterative_add_se(sedict, chrom, group, se_ordered[0])
+	sedict, added = iterative_add_se(sedict=sedict, chrom=chrom, group=group, se=se_ordered[0], 
+				  all_se_by_chrom=all_se_by_chrom, window=window, isbed=isbed)
 	for se in se_ordered[1:]:
 		overlapped_loci = []
 		#overlapped_intervals = [] # unused
@@ -251,16 +177,18 @@ def run_iterative_add_se(chrom):  # add single exon genes iteratively, assumes t
 		added = False
 		for loci in overlapped_loci:
 			g = loci[0]
-			sedict, added = iterative_add_se(sedict, chrom, g, se)
+			sedict, added = iterative_add_se(sedict=sedict, chrom=chrom, group=g, se=se, 
+				  all_se_by_chrom=all_se_by_chrom, window=window, isbed=isbed)
 			if added:
 				break
 		if not added:
 			group += 1
-			sedict, added = iterative_add_se(sedict, chrom, group, se)
+			sedict, added = iterative_add_se(sedict=sedict, chrom=chrom, group=group, se=se, 
+				  all_se_by_chrom=all_se_by_chrom, window=window, isbed=isbed)
 	return sedict
 
 
-def find_best_tss(sites, finding_tss, remove_used):
+def find_best_tss(sites, finding_tss, remove_used, window):
 	nearby = dict.fromkeys(sites, 0)  # key: site, value: number of supporting reads in window
 	wnearby = dict.fromkeys(sites, 0)  # key: site, value: weighted number of supporting reads in window
 	bestsite = [0, 0, 0, 0]  # TSS position, wnearby value, nearby value, number of supporting reads
@@ -278,15 +206,14 @@ def find_best_tss(sites, finding_tss, remove_used):
 				nearby[s] += sites[s_]
 		if wnearby[s] > bestsite[1]:  # update bestsite if this site has more supporting reads
 			bestsite = [s, wnearby[s], nearby[s], sites[s]]
-
+	# TODO: remove used is always set to True, this if statement is unneccesary
 	if remove_used:
 		for s in list(sites.keys()):  # remove reads supporting the bestsite to find alternative TSSs
 			if abs(s - bestsite[0]) <= window:
 				sites.pop(s)
 	return sites, bestsite
 
-
-def find_tsss(sites, total, finding_tss=True, max_results=2, chrom='', junccoord='', remove_used=True):
+def find_tsss(sites, total, window, annot_tss=False, annot_tes=False, minsupport=0.25, no_redundant='none', finding_tss=True, max_results=2, chrom='', junccoord='', remove_used=True):
 	""" Finds the best TSSs within Sites. If find_tss is False, some
 	assumptions are changed to search specifically for TESs. I also assume that the correct
 	splice site will be the more represented, so measures to filter out degraded reads are
@@ -297,16 +224,16 @@ def find_tsss(sites, total, finding_tss=True, max_results=2, chrom='', junccoord
 	avg = remaining / (len(sites))
 	while ((minsupport < 1 and remaining/total > minsupport) or remaining >= minsupport) and \
 			len(found_tss) < max_results:
-		sites, bestsite = find_best_tss(sites, finding_tss, remove_used)
+		sites, bestsite = find_best_tss(sites, finding_tss, remove_used, window)
 		newremaining = sum(list(sites.values()))
 		used = remaining - newremaining
 		remaining = newremaining
-		if len(found_tss) >= 1 and (args.n == 'best_only' or
+		if len(found_tss) >= 1 and (no_redundant == 'best_only' or
 			(minsupport < 1 and (used/total) < minsupport or bestsite[3] < 4 or used < avg)):
 			# second+ end site called stringently
 			break
 		closest_annotated = 1e15  # just a large number
-		if annot_tss and chrom in annot_tss:  # args.f supplied
+		if annot_tss and chrom in annot_tss:  # args.gtf supplied
 			for t in range(bestsite[0]-window, bestsite[0]+window):
 				if (finding_tss and t in annot_tss[chrom] or not finding_tss and t in annot_tes[chrom]) \
 					and abs(t - bestsite[0]) < abs(closest_annotated - bestsite[0]) and \
@@ -321,7 +248,8 @@ def find_tsss(sites, total, finding_tss=True, max_results=2, chrom='', junccoord
 	return found_tss
 
 
-def find_best_sites(sites_tss_all, sites_tes_all, junccoord, chrom='', max_results=max_results):
+def find_best_sites(sites_tss_all, sites_tes_all, junccoord, no_redundant, annot_tss=False, annot_tes=False, 
+		    minsupport=0.25, chrom='', max_results=2, window=100):
 	""" sites_tss_all = {tss: count}
 	sites_tes_all = {tss: {tes: count}}
 	specific_tes = {tes: count} for a specific set of tss within given window
@@ -330,7 +258,8 @@ def find_best_sites(sites_tss_all, sites_tes_all, junccoord, chrom='', max_resul
 	# if junccoord in [(36178823, 36180248), (36179025, 36180248)]:
 	# 	print('--', junccoord)
 	# 	print(sites_tss_all)
-	found_tss = find_tsss(sites_tss_all, total, finding_tss=True, max_results=max_results,
+	found_tss = find_tsss(sites= sites_tss_all, total=total, annot_tss=annot_tss, annot_tes=annot_tes, no_redundant=no_redundant,
+		        window=window, minsupport=minsupport, finding_tss=True, max_results=max_results,
 		chrom=chrom, junccoord=junccoord)
 	# if junccoord in [(36178823, 36180248), (36179025, 36180248)]:
 	# 	print(found_tss)
@@ -347,14 +276,15 @@ def find_best_sites(sites_tss_all, sites_tes_all, junccoord, chrom='', max_resul
 						specific_tes[tes] = 0
 					specific_tes[tes] += sites_tes_all[tss_][tes]
 
-		found_tes = find_tsss(specific_tes, total, finding_tss=False, max_results=max_results, chrom=chrom,
+		found_tes = find_tsss(sites=specific_tes, total=total, window=window, finding_tss=False, max_results=max_results, chrom=chrom,
 			junccoord=junccoord)
+
 		for tes in found_tes:
 			ends += [[tss[0], tes[0], max(tss[3], tes[3]), tss[3], tes[3]]]
 	return ends
 
 
-def run_se_collapse(chrom):
+def run_se_collapse(chrom, singleexon, isbed, clean, no_redundant, isoformtss, minsupport, annot_tss, window):
 	senames = {}
 	towrite = []
 	used_ends = {}
@@ -363,9 +293,12 @@ def run_se_collapse(chrom):
 	for locus in singleexon[chrom]:
 		line = list(singleexon[chrom][locus]['line'])
 		locus_info = singleexon[chrom][locus]
-		ends = find_best_sites(locus_info['tss'], locus_info['tss_tes'],
-			locus_info['bounds'], chrom, max_results=1)
-		name = line[3] if bed else line[9]
+
+		ends = find_best_sites(sites_tss_all=locus_info['tss'], sites_tes_all=locus_info['tss_tes'], 
+			 junccoord=locus_info['bounds'], chrom=chrom, 
+			no_redundant=no_redundant, annot_tss=annot_tss, minsupport=minsupport, 
+			max_results=1, window=window)
+		name = line[3] if isbed else line[9]
 		if name not in senames:
 			senames[name] = 0
 		strand = sorted(singleexon[chrom][locus]['strand'].items(),key=lambda x: x[1])[-1][1]
@@ -373,7 +306,7 @@ def run_se_collapse(chrom):
 			if (tss, tes) not in used_ends:
 				used_ends[(tss, tes)] = len(towrite)
 			else:
-				if not args.c:
+				if not clean:
 					towrite[used_ends[(tss, tes)]][-1] += support
 				continue
 			if tss not in all_se_starts:
@@ -384,33 +317,33 @@ def run_se_collapse(chrom):
 			all_se_ends[tes] += support
 
 			i = senames[name]
-			if not bed:
+			if not isbed:
 				edited_line = edit_line(line, tss, tes, tes-tss)
 				edited_line[5] = strand
 			else:
 				edited_line = edit_line_bed12(line, tss, tes, tes-tss)
 				edited_line[8] = strand
-			if not args.c:
+			if not clean:
 				edited_line += [support]
 			if i >= 1:  # to avoid redundant names for isoforms from the same general locus
 				if '_' in name:
 					newname = name[:name.rfind('_')]+'-'+str(i)+name[name.rfind('_'):]
 				else:
 					newname = name+'-'+str(i)
-				if not bed:
+				if not isbed:
 					edited_line[9] = newname
 				else:
 					edited_line[3] = newname
 			senames[name] += 1
 			towrite += [edited_line]
 
-	if args.i:
+	if isoformtss:
 		return towrite
 	new_towrite = []
 	used_ends = {}
 	for line in towrite:
 		line = list(line)
-		if bed:
+		if isbed:
 			tss, tes = int(line[1]), int(line[2])
 		else:
 			tss, tes = int(line[15]), int(line[16])
@@ -429,10 +362,10 @@ def run_se_collapse(chrom):
 		if (tss, tes) not in used_ends:
 			used_ends[(tss, tes)] = len(new_towrite)
 		else:
-			if not args.c:
+			if not clean:
 				new_towrite[used_ends[(tss, tes)]][-1] += line[-1]  # support
 			continue
-		if bed:
+		if isbed:
 			newline = edit_line_bed12(line, tss, tes)
 		else:
 			newline = edit_line(line, tss, tes)
@@ -440,7 +373,8 @@ def run_se_collapse(chrom):
 	return new_towrite
 
 
-def run_find_best_sites(chrom):
+# this function can only have one input because it is run with map
+def run_find_best_sites(chrom, isoforms, isoformtss, clean, isbed, no_redundant, window, annot_tss, annot_tes, minsupport):
 	allends = {}  # counts of all TSSs/TESs by chromosome
 	allends[chrom] = {}
 	allends[chrom]['tss'] = {}
@@ -451,49 +385,51 @@ def run_find_best_sites(chrom):
 		towrite[chrom][jset] = []
 		line = list(isoforms[chrom][jset]['line'])
 		junccoord = isoforms[chrom][jset]['junccoord']
-		jset_ends = find_best_sites(isoforms[chrom][jset]['tss'],
-			isoforms[chrom][jset]['tss_tes'], junccoord, chrom)
+		jset_ends = find_best_sites(sites_tss_all=isoforms[chrom][jset]['tss'],
+			sites_tes_all=isoforms[chrom][jset]['tss_tes'], junccoord=junccoord, chrom=chrom, 
+			no_redundant=no_redundant, annot_tss=annot_tss, annot_tes=annot_tes, 
+		    minsupport=minsupport,  window=window)
 
-		if args.i and args.n == 'longest':
+		if isoformtss and no_redundant == 'longest':
 			tss_sorted = sorted(jset_ends, key=lambda x: x[0])[0]  # smallest left coord
 			tes_sorted = sorted(jset_ends, key=lambda x: x[1])[-1]  # largest right coord
-			if not args.c:
+			if not clean:
 				line += [(tss_sorted[3]+tes_sorted[3])/2.]
 			tss, tes = tss_sorted[0], tes_sorted[1]
-			if not bed:
+			if not isbed:
 				towrite[chrom][jset] += [edit_line(line, tss, tes)]
 			else:
 				towrite[chrom][jset] += [edit_line_bed12(line, tss, tes)]
 			continue
-		if args.i and args.n == 'best_only':
+		if isoformtss and no_redundant == 'best_only':
 			tss, tes = jset_ends[0][:2]
-			if not args.c:
+			if not clean:
 				line += [jset_ends[0][3]]
-			if not bed:
+			if not isbed:
 				towrite[chrom][jset] += [edit_line(line, tss, tes)]
 			else:
 				towrite[chrom][jset] += [edit_line_bed12(line, tss, tes)]
 			continue  # 1 isoform per unique set of junctions
 
 		i = 0
-		name = line[9] if not bed else line[3]
+		name = line[9] if not isbed else line[3]
 		for tss, tes, support, tsscount, tescount in jset_ends:
-			if not bed:
+			if not isbed:
 				edited_line = edit_line(line, tss, tes)
 			else:
 				edited_line = edit_line_bed12(line, tss, tes)
-			if not args.c:
+			if not clean:
 				edited_line += [support]
 			if i >= 1:  # to avoid redundant names for isoforms with the same junctions
 				if '_' in name:
 					newname = name[:name.rfind('_')]+'-'+str(i)+name[name.rfind('_'):]
 				else:
 					newname = name+'-'+str(i)
-				if not bed:
+				if not isbed:
 					edited_line[9] = newname
 				else:
 					edited_line[3] = newname
-			if args.i:
+			if isoformtss:
 				towrite[chrom][jset] += [edited_line]
 			else:  # all isoforms will go through another pass to homogenize ends
 				towrite[chrom][jset] += [edited_line + [junccoord]]
@@ -503,9 +439,9 @@ def run_find_best_sites(chrom):
 				if tes not in allends[chrom]['tes']:
 					allends[chrom]['tes'][tes] = 0
 				allends[chrom]['tes'][tes] = tescount
-			if args.n != 'longest':
+			if no_redundant != 'longest':
 				i += 1
-	if args.i:
+	if isoformtss:
 		return towrite
 
 	new_towrite = {}  # another pass through all isoforms, moving tss/tes to be more uniform within a gene
@@ -517,7 +453,7 @@ def run_find_best_sites(chrom):
 		for line in towrite[chrom][jset]:  # adjust isoform TSS/TES using allends dict
 			line = list(line)
 			junccoord = line[-1]
-			if bed:
+			if isbed:
 				tss, tes = int(line[1]), int(line[2])
 			else:
 				tss, tes = int(line[15]), int(line[16])
@@ -541,19 +477,19 @@ def run_find_best_sites(chrom):
 
 			if (tss,tes) not in jsetends:
 				jsetends.add((tss,tes))
-				if args.n == 'best_only':
+				if no_redundant == 'best_only':
 					endpair = [tss, tes, support] if endpair[2] < tss_support else endpair
-				elif args.n == 'longest':
+				elif no_redundant == 'longest':
 					endpair[0] = tss if tss < endpair[0] else endpair[0]
 					endpair[1] = tes if tes > endpair[1] else endpair[1]
 				else:
-					if bed:
+					if isbed:
 						newline = edit_line_bed12(line, tss, tes)
 					else:
 						newline = edit_line(line, tss, tes)
 					new_towrite[chrom][jset] += [newline[:-1]]
 		if endpair[0] != 1e15:  # write best_only or longest option isoforms
-			if bed:
+			if isbed:
 				newline = edit_line_bed12(line, endpair[0], endpair[1])
 			else:
 				newline = edit_line(line, endpair[0], endpair[1])
@@ -582,7 +518,6 @@ def edit_line(line, tss, tes, blocksize=''):
 	line[20] = ','.join([str(x) for x in bstarts])+','
 	return line
 
-
 def edit_line_bed12(line, tss, tes, blocksize=''):
 	line = list(line)
 	if blocksize:
@@ -605,112 +540,140 @@ def edit_line_bed12(line, tss, tes, blocksize=''):
 	line[11] = ','.join([str(x - int(line[1])) for x in bstarts])+','
 	return line
 
+def collapse_isoforms_precise(queryfile, max_results=2, window=100, threads=2, clean=False,
+		minsupport=0.25, gtfname=None, no_redundant='none', nosplice='chrM', isoformtss=False, 
+		outputfname=None, isbed=True, quiet=False):
+	if outputfname:
+		if outputfname[-3:].lower() != queryfile[-3:].lower():
+			sys.stderr.write('Make sure input and output file extensions agree\n')
+			sys.exit(1)
+	else:  # default output name
+		outputfile = queryfile[:-3]+'collapsed.bed' if isbed else queryfile[:-3]+'collapsed.psl'
 
-annot_tss = {}  # annotated left terminal sites per chromosome from GTF
-annot_tes = {}  # annotated right terminal sites
-if args.f:
-	for line in gtf:
-		if line.startswith('#'):
+	# get annotation starts and ends if an input gtf has been provided
+	annot_tss = {}  # annotated left terminal sites per chromosome from GTF
+	annot_tes = {}  # annotated right terminal sites
+	if gtfname:
+		gtf = open(gtfname, 'r')
+		for line in gtf:
+			if line.startswith('#'):
+				continue
+			line = line.rstrip().split('\t')
+			chrom, ty, start, end = line[0], line[2], int(line[3]) - 1, int(line[4])
+			if ty == 'transcript':
+				if chrom not in annot_tss:
+					annot_tss[chrom] = set()
+					annot_tes[chrom] = set()
+				annot_tss[chrom].add(start)
+				annot_tes[chrom].add(end)
+		if not quiet:
+			sys.stderr.write('Annotated ends extracted from GTF\n')
+
+	isoforms = {}  # spliced isoforms
+	all_se_by_chrom = {}  # single-exon reads grouped by exact start and end site
+	nosplice_chroms = set(nosplice.split(','))
+	query = open(queryfile, 'r')
+	for line in query:
+		line = tuple(line.rstrip().split('\t'))
+		if isbed:
+			chrom, tss, tes = line[0], int(line[1]), int(line[2])
+			junctions, junccoord = get_junctions_bed12(line)
+		else:
+			chrom, tss, tes = line[13], int(line[15]), int(line[16])
+			junctions, junccoord = get_junctions(line)
+
+		if not junctions:  # single-exon isoforms
+			if chrom not in all_se_by_chrom:
+				all_se_by_chrom[chrom] = {}
+			if (tss, tes) not in all_se_by_chrom[chrom]:
+				all_se_by_chrom[chrom][(tss, tes)] = {}
+				all_se_by_chrom[chrom][(tss, tes)]['count'] = 0
+				all_se_by_chrom[chrom][(tss, tes)]['line'] = line
+			all_se_by_chrom[chrom][(tss, tes)]['count'] += 1
 			continue
-		line = line.rstrip().split('\t')
-		chrom, ty, start, end = line[0], line[2], int(line[3]) - 1, int(line[4])
-		if ty == 'transcript':
-			if chrom not in annot_tss:
-				annot_tss[chrom] = set()
-				annot_tes[chrom] = set()
-			annot_tss[chrom].add(start)
-			annot_tes[chrom].add(end)
-	if not args.quiet:
-		sys.stderr.write('Annotated ends extracted from GTF\n')
+		elif chrom in nosplice_chroms:
+			continue
 
-isoforms = {}  # spliced isoforms
-all_se_by_chrom = {}  # single-exon reads grouped by exact start and end site
-nosplice_chroms = set(args.nosplice.split(','))
-for line in psl:
-	line = tuple(line.rstrip().split('\t'))
-	if bed:
-		chrom, tss, tes = line[0], int(line[1]), int(line[2])
-		junctions, junccoord = get_junctions_bed12(line)
-	else:
-		chrom, tss, tes = line[13], int(line[15]), int(line[16])
-		junctions, junccoord = get_junctions(line)
+		junctions = str(sorted(list(junctions)))  # splice junction chain
 
-	if not junctions:  # single-exon isoforms
-		if chrom not in all_se_by_chrom:
-			all_se_by_chrom[chrom] = {}
-		if (tss, tes) not in all_se_by_chrom[chrom]:
-			all_se_by_chrom[chrom][(tss, tes)] = {}
-			all_se_by_chrom[chrom][(tss, tes)]['count'] = 0
-			all_se_by_chrom[chrom][(tss, tes)]['line'] = line
-		all_se_by_chrom[chrom][(tss, tes)]['count'] += 1
-		continue
-	elif chrom in nosplice_chroms:
-		continue
+		if chrom not in isoforms:
+			isoforms[chrom] = {}
+		if junctions not in isoforms[chrom]:
+			isoforms[chrom][junctions] = {}
+			isoforms[chrom][junctions]['tss'] = {}  # ends stored for each unique sjc
+			isoforms[chrom][junctions]['tss_tes'] = {}
+			isoforms[chrom][junctions]['line'] = line
+			isoforms[chrom][junctions]['junccoord'] = junccoord
 
-	junctions = str(sorted(list(junctions)))  # splice junction chain
-	# if '(36179025, 36179262)' in junctions:
-	# 	print(tss, junccoord, line[3])
+		if tss not in isoforms[chrom][junctions]['tss']:
+			isoforms[chrom][junctions]['tss'][tss] = 0  # tss usage count
+			isoforms[chrom][junctions]['tss_tes'][tss] = {}  # TESs for this TSS
+		isoforms[chrom][junctions]['tss'][tss] += 1
 
-	if chrom not in isoforms:
-		isoforms[chrom] = {}
-	if junctions not in isoforms[chrom]:
-		isoforms[chrom][junctions] = {}
-		isoforms[chrom][junctions]['tss'] = {}  # ends stored for each unique sjc
-		isoforms[chrom][junctions]['tss_tes'] = {}
-		isoforms[chrom][junctions]['line'] = line
-		isoforms[chrom][junctions]['junccoord'] = junccoord
+		if tes not in isoforms[chrom][junctions]['tss_tes'][tss]:
+			isoforms[chrom][junctions]['tss_tes'][tss][tes] = 0
+		isoforms[chrom][junctions]['tss_tes'][tss][tes] += 1
 
-	if tss not in isoforms[chrom][junctions]['tss']:
-		isoforms[chrom][junctions]['tss'][tss] = 0  # tss usage count
-		isoforms[chrom][junctions]['tss_tes'][tss] = {}  # TESs for this TSS
-	isoforms[chrom][junctions]['tss'][tss] += 1
+	if not quiet:
+		sys.stderr.write('Read data extracted\n')
 
-	if tes not in isoforms[chrom][junctions]['tss_tes'][tss]:
-		isoforms[chrom][junctions]['tss_tes'][tss][tes] = 0
-	isoforms[chrom][junctions]['tss_tes'][tss][tes] += 1
+	chrom_names = []  # sorted by descending total number of unique single-exon isoforms
+	for chrom in all_se_by_chrom:
+		chrom_names += [(chrom, len(all_se_by_chrom[chrom]))]
+	chrom_names = [chrom for chrom,num in sorted(chrom_names, key=lambda x: x[1], reverse=True)]
 
-if not args.quiet:
-	sys.stderr.write('Read data extracted\n')
+	# Add single exon genes iteratively by chromosome
+	# partial fills in all named elements except chrom. 
+	# when called with a single argument, it's automatically assigned to the empty one
+	# this way we can run map with one argument and we don't need to give it a list of lists (of unnamed arguments)
+	named_run_iterative_add_se = partial(run_iterative_add_se, all_se_by_chrom=all_se_by_chrom, 
+				      window=window, isbed=isbed)
 
-chrom_names = []  # sorted by descending total number of unique single-exon isoforms
-for chrom in all_se_by_chrom:
-	chrom_names += [(chrom, len(all_se_by_chrom[chrom]))]
-chrom_names = [chrom for chrom,num in sorted(chrom_names, key=lambda x: x[1], reverse=True)]
-res = {}
-if __name__ == '__main__':
-	p = Pool(args.t)
-	res = p.map(run_iterative_add_se, chrom_names)
-	p.terminate()
-all_se_by_chrom = None
-
-singleexon = {}  # single-exon isoforms
-for r in res:  # combine results
-	singleexon.update(r)
-
-if not args.quiet:
-	sys.stderr.write('Single-exon genes grouped, collapsing\n')
-
-with open(args.o, 'wt') as outfile:
-	writer = csv.writer(outfile, delimiter='\t', lineterminator=os.linesep)
-
+	res = {}
 	if __name__ == '__main__':
-		p = Pool(args.t)
-		res = p.map(run_se_collapse, chrom_names)
+		p = Pool(threads)
+		res = p.map(named_run_iterative_add_se, chrom_names)
 		p.terminate()
-	singleexon = None
+	all_se_by_chrom = None
 
-	for r in res:
-		for edited_line in r:
-			writer.writerow(edited_line)
+	singleexon = {}  # single-exon isoforms
+	for r in res:  # combine results
+		singleexon.update(r)
 
-	if __name__ == '__main__':
-		p = Pool(args.t)
-		res = p.map(run_find_best_sites, list(isoforms.keys()))
+	if not quiet:
+		sys.stderr.write('Single-exon genes grouped, collapsing\n')
+
+	with open(outputfile, 'wt') as outfile:
+		writer = csv.writer(outfile, delimiter='\t', lineterminator=os.linesep)
+
+    # SE collapse
+		named_se_input = partial(run_se_collapse, singleexon=singleexon, isbed=isbed, 
+			clean=clean, no_redundant=no_redundant, isoformtss=isoformtss, 
+			minsupport=minsupport, annot_tss=annot_tss, window=window)
+#		if __name__ == '__main__': #???
+		p = Pool(threads)
+		res = p.map(named_se_input, chrom_names)
 		p.terminate()
-	isoforms = None
+		singleexon = None
 
-	for towrite in res:
-		for chrom in towrite:
-			for jset in towrite[chrom]:
-				for iso in towrite[chrom][jset]:
-					writer.writerow(iso)
+		for r in res:
+			for edited_line in r:
+				writer.writerow(edited_line)
+
+		named_run_find_best_sites = partial(run_find_best_sites, isoforms=isoforms, isoformtss=isoformtss, 
+				     clean=clean, isbed=isbed, no_redundant=no_redundant, window=window, annot_tss=annot_tss,
+						annot_tes=annot_tes, minsupport=minsupport)
+#		if __name__ == '__main__':
+		p = Pool(threads)
+		res = p.map(named_run_find_best_sites, list(isoforms.keys()))
+		p.terminate()
+		isoforms = None
+
+		for towrite in res:
+			for chrom in towrite:
+				for jset in towrite[chrom]:
+					for iso in towrite[chrom][jset]:
+						writer.writerow(iso)
+
+if __name__ == "__main__":
+    main()
