+++ src/flair/select_from_bed.py
@@ -2,16 +2,18 @@
 import sys
 import csv
 import os
-keepnames = set()
 
-for line in open(sys.argv[1]):  # bed of promoter-supported reads
-	line = line.rstrip().split('\t')
-	keepnames.add(line[3])
+def select_from_bed(inputbed, queryfile, outputfile, isbed=True):
+	keepnames = set()
 
-with open(sys.argv[3], 'wt') as outfile:
-	writer = csv.writer(outfile, delimiter='\t', lineterminator=os.linesep)
-	isbed = sys.argv[2][-3:].lower() != 'psl'
-	for line in open(sys.argv[2]):  # psl or bed
+	for line in open(inputbed):  # bed of promoter-supported reads
 		line = line.rstrip().split('\t')
-		if isbed and line[3] in keepnames or not isbed and line[9] in keepnames:
-			writer.writerow(line)
+		keepnames.add(line[3])
+
+	with open(outputfile, 'wt') as outfile:
+		writer = csv.writer(outfile, delimiter='\t', lineterminator=os.linesep)
+		for line in open(queryfile):  # psl or bed
+			line = line.rstrip().split('\t')
+			if isbed and line[3] in keepnames or not isbed and line[9] in keepnames:
+				writer.writerow(line)
+
